Tests et modifications de la méthode de rééchantillonage + adaptation de la classe CSV

+++ include/csv.h
@@ -5,7 +5,7 @@ class CSV
 {
 public:
     CSV();
-    float **readCSV(char *);
+    double **readCSV(char *);
 
     int getNbLines();
